third_party/blink/web_tests/platform/generic/svg/animations/animate-end-attribute-expected.txt

SVG 1.1 dynamic animation tests

Tests end conditions are respected properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 50
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 200
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 200
PASS rect.x.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
